Moved service related classes to svc package, updated some classes to public
