+++ backend/src/main/java/com/empire/Player.java
@@ -6,7 +6,7 @@ import com.google.appengine.api.datastore.EntityNotFoundException;
 import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.KeyFactory;
 
-final class Player {
+public final class Player {
 	public static String TYPE = "Player";
 
 	public final String email;
